# perf: improve perf of key order