--- a/lib/document.js
+++ b/lib/document.js
@@ -1386,6 +1386,7 @@ Document.prototype.$set = function $set(path, val, type, options) {
     }
 
     let popOpts;
+    const typeKey = this.$__schema.options.typeKey;
     if (schema.options &&
         Array.isArray(schema.options[typeKey]) &&
         schema.options[typeKey].length &&