I actually dont understand why the key ordre is so important. Nevertheless, this implementatoin is faster.

without reordering:
MONGOOSE_DEV=1 node create.js 
CheckItem x 201,161 ops/sec ±1.00% (92 runs sampled)
Board x 182 ops/sec ±0.82% (83 runs sampled)

before:
 MONGOOSE_DEV=1 node create.js 
CheckItem x 170,092 ops/sec ±1.35% (90 runs sampled)
Board x 168 ops/sec ±1.22% (80 runs sampled)

after:
MONGOOSE_DEV=1 node create.js 
CheckItem x 195,166 ops/sec ±0.41% (96 runs sampled)
Board x 174 ops/sec ±1.29% (80 runs sampled)

So we get pretty close to the non-ordered object.