@Uzlopak 2022-04-08
You may ask why this is faster, and maybe it is even more correct.

So in the original solution, the keys are deleted and added back to the object, as such the keys will be added at the end of the Object and thus sorting the keys. This of course works, but if you have by any chance some additional keys, which you were not aware of, they will be at the beginning.

E.g.

you have a object with following keys

```
firstName
lastName
_id
additional
```

and we have to sort them by the schema and the schema has following keys:
```
_id
firstName
lastName
```

then with the old implementation, you would get 

```
additional
_id
firstName
lastName
```

The new way makes use of the fact, that if you assign an Object to an other Object, it will use the order of the keys of the other Object. So we first generate a new Object with keys where the value is undefined. Then we assign the original Object to the new Object and voila have all the keys in the right order. Additional keys will be at the end of the object.

Like this:

```
_id
firstName
lastName
additional
```