@Uzlopak 2022-04-08
@AbdelrahmanHafez 

The unit test for this is already existing.

https://github.com/Automattic/mongoose/blob/dc0a332d5494501725c542a3be9762a5b9093d87/test/document.test.js#L10556